requirements.txt
flask>=3.0,<4
google-genai>=1.13.0
